Split out MP4 functions into a class
Moving find_atom into a class with a file pointer accessible as a data
member exposed a bug where we were seeking beyond the end of the file
and not triggering EOF, which would sometimes cause reads to endlessly
get '0' and result in an infinite loop. Fixed by adding an feof() check
after every read in find_atom.