Improve ID3v2 search
Increase the search field, as frames may not appear for many thousands
of bytes. If a frame is not found, return to the start in case a
different frame was within the field. Remove fallbacks to ID3v1 since
we should be able to be certain of what kind of file we're looking at
and we're no longer set up to seek to the end of the file when falling
back to ID3v1.